chore: Migrate to Rails (Omakase) style
We used to use Shopify style.  Let's use a Rails style which will
probably become more widespread in the future.

Note that the only exclusion is Rails/RefuteMethods.  It mandates us to
inherit from ActiveSupport::TestCase which I prefer not to use at the
moment to aviod ActiveSupport monkey patching side effects.